this is Main
Just updated some of the info on this file in case someone wants to mess around with it and doesn't want to read each file to figure out what the functions are doing in this section.  Or at least, I attempted to clarify some of whats going on.  That someone can mess around after all the files are linked up in a project